Refactor streaming methods in EngineCloudFlare
The code has been modified to refactor the streaming methods in the EngineCloudFlare class. Substantial changes are in the `stream_generate` and `async_stream_generate` methods where we've removed the check for the '[DONE]' event and instead directly yield instances of `ChatCompletionModel` if the event isn't '[DONE]'. This simplifies the code and makes it more readable.